pyproject.toml
[project]
name = "cli-agent"
version = "0.1.0"
description = "A simple CLI coding agent using Google's Gemini API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
